pyproject.toml
[project]
name = "media-alchemy"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cohere>=5.13.12",
    "ctc-segmentation>=1.7.4",
    "deepmultilingualpunctuation>=1.0.1",
    "faster-whisper>=1.1.1",
    "ffmpeg-python>=0.2.0",
    "gradio>=5.16.1",
    "langchain>=0.3.19",
    "langchain-cohere>=0.4.2",
    "langchain-experimental>=0.3.4",
    "langchain-neo4j>=0.3.0",
    "langchain-openai>=0.3.6",
    "neo4j>=5.28.1",
    "neo4j-graphrag>=1.5.0",
    "numpy<2",
    "openai>=1.63.2",
    "pyannote-audio>=3.3.2",
    "python-dotenv>=1.0.1",
    "reportlab>=4.3.1",
    "sentence-transformers>=3.4.1",
    "transformers>=4.49.0",
    "triton>=3.2.0",
    "yt-dlp>=2025.1.26",
]
